Renamed sessions.py to game_session.py
